[audio] Refactor to useEvent
