+++ apps/native-component-list/src/screens/Audio/AV/AudioScreen.tsx
@@ -5,8 +5,8 @@ import { PixelRatio, ScrollView, StyleSheet } from 'react-native';
 
 import AudioModeSelector from './AudioModeSelector';
 import Player from './AudioPlayer';
-import HeadingText from '../../components/HeadingText';
-import ListButton from '../../components/ListButton';
+import HeadingText from '../../../components/HeadingText';
+import ListButton from '../../../components/ListButton';
 
 export default class AudioScreen extends React.Component {
   static navigationOptions = {
@@ -32,7 +32,7 @@ export default class AudioScreen extends React.Component {
         />
         <HeadingText>Local asset player</HeadingText>
         <Player
-          source={Asset.fromModule(require('../../../assets/sounds/polonez.mp3'))}
+          source={Asset.fromModule(require('../../../../assets/sounds/polonez.mp3'))}
           style={styles.player}
         />
       </ScrollView>
