+++ apps/native-component-list/src/screens/Audio/expo-audio/Recorder.tsx
@@ -13,7 +13,7 @@ import {
 } from 'react-native';
 
 import AudioInputSelector from './AudioInputSelector';
-import Colors from '../../constants/Colors';
+import Colors from '../../../constants/Colors';
 
 type RecorderProps = {
   onDone?: (uri: string) => void;
