+++ apps/native-component-list/src/screens/ExpoApisScreen.tsx
@@ -29,8 +29,7 @@ const screens = [
   'Alert',
   'Appearance',
   'AppleAuthentication',
-  'Audio (expo-av)',
-  'Audio (expo-audio)',
+  'Audio',
   'AsyncStorage',
   'AuthSession',
   'BackgroundFetch',
@@ -69,8 +68,6 @@ const screens = [
   'Permissions',
   'Print',
   'Random',
-  'Recording (expo-av)',
-  'Recording (expo-audio)',
   'SMS',
   'NavigationBar',
   'SafeAreaContext',
