+++ packages/expo-audio/src/Audio.types.ts
@@ -17,19 +17,6 @@ export type RecordingInput = {
   uid: string;
 };
 
-export type AudioPlayerState = {
-  isLoaded: boolean;
-  isLooping: boolean;
-  isMuted: boolean;
-  positionMillis: number;
-  durationMillis: number;
-  rate: number;
-  volume: number;
-  playing: boolean;
-  audioPan: number;
-  shouldCorrectPitch: boolean;
-};
-
 export type PitchCorrectionQuality = 'low' | 'medium' | 'high';
 
 export type AudioStatus = {
